+++ addOns/spiderAjax/CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## Unreleased
 ### Added
 - Allow to configure how the scope is checked, either Flexible or Strict, to allow or not access to out of scope domains.
+- Allow to avoid logout elements.
 
 ### Changed
 - Maintenance changes.